feat: improve aspen-core API ergonomics and validation
Add builder methods to WriteRequest/DeleteRequest for more ergonomic API usage:
- WriteRequest::set(), set_with_ttl(), delete(), compare_and_swap()
- DeleteRequest::new()

Document KeyValueWithRevision field semantics:
- Explain version, create_revision, mod_revision fields
- Detail MVCC patterns and optimistic concurrency control

Fix usize/u32 type consistency in error fields:
- Change size fields to u32 in KeyTooLarge/ValueTooLarge/BatchTooLarge
- Update all validation callsites and tests

Add InitRequest validation:
- Check for empty members, node ID 0, duplicate IDs
- Consolidate validation logic in InitRequest::validate()

All changes maintain API compatibility and pass existing tests.

+++ crates/aspen-core/src/error.rs
@@ -75,7 +75,7 @@ pub enum KeyValueStoreError {
     #[error("key size {size} exceeds maximum of {max} bytes")]
     KeyTooLarge {
         /// Actual size of the key in bytes.
-        size: usize,
+        size: u32,
         /// Maximum allowed size in bytes.
         max: u32,
     },
@@ -84,7 +84,7 @@ pub enum KeyValueStoreError {
     #[error("value size {size} exceeds maximum of {max} bytes")]
     ValueTooLarge {
         /// Actual size of the value in bytes.
-        size: usize,
+        size: u32,
         /// Maximum allowed size in bytes.
         max: u32,
     },
@@ -93,7 +93,7 @@ pub enum KeyValueStoreError {
     #[error("batch size {size} exceeds maximum of {max} keys")]
     BatchTooLarge {
         /// Actual number of keys in the batch.
-        size: usize,
+        size: u32,
         /// Maximum allowed batch size.
         max: u32,
     },
